+++ PlexMatchGenerator/Helpers/ArgumentHelper.cs
@@ -6,7 +6,7 @@ namespace PlexMatchGenerator.Helpers
 {
     public class ArgumentHelper
     {
-        public static GeneratorOptions ProcessCommandLineResults(string plexToken, string plexUrl, List<string> rootPaths, string logPath)
+        public static GeneratorOptions ProcessCommandLineResults(string plexToken, string plexUrl, List<string> rootPaths, string logPath, bool noOverwrite)
         {
             //ensure we end the path with a slash
             if (logPath != null && !logPath.EndsWith("\\") && !logPath.EndsWith('/'))
@@ -22,7 +22,8 @@ namespace PlexMatchGenerator.Helpers
                 LogPath = logPath,
                 PlexServerUrl = plexUrl,
                 PlexServerToken = plexToken,
-                RootPaths = GenerateRootPaths(rootPaths)
+                RootPaths = GenerateRootPaths(rootPaths),
+                NoOverwrite = noOverwrite
             };
         }