+++ .github/workflows/ReleaseNotes.md
@@ -1,7 +1,5 @@
 ## What's Changed
-* Build pipeline now includes bugfix branches
-* Fixed a bug throwing a Null Reference Exception when optional parameter -r was not used
-* Fixed a bug that would write to the incorrect file when log parameter did not include trailing slash
+* Add command switch to disable automatic overwriting as requested in [#30](https://github.com/johnkiddjr/PlexMatch-File-Generator/issues/30)
 
 ## Known Bugs
 * None! If you encounter any bugs, please open an [issue](https://github.com/johnkiddjr/PlexMatch-File-Generator/issues/new)!